sisusf_app/requirements.txt
# Bibliotecas necessárias

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug==3.0.1